% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{package}
\name{analyzeJ2Kmodel}
\alias{analyzeJ2Kmodel}
\title{analyzeJ2Kmodel Package}
\description{
This package is created exclusively for testing the basics details while setting up the JAMS/J2000 hydrological model (http:\\jams.uni-jena.de\\) in a new catchment. Currently it is in the beta mode, therefore, comments and suggestions are are highly appreciated.
}
\details{
Currently, some considerations are to be made for running the package efficiently. Firstly, make "The path where the jam file is situated" as the "workingFolder" for running the package. For example: if your jam file of the model is in the following location: "C:\\J2K_model\\J2K_koshi_modified.jam", then you have to write: "workingFolder" <- "C:\\J2K_model\\". Based on this consideration all the calulations are made. 

Also, make sure to install the other packages required to run this package, by yourself. You can find the readme file in the following link:  "https:\\github.com\\kabiraj404\\analyzeJ2Kmodel\\blob\\master\\README.md"
}
\author{
Kabiraj \email{kabiraj(dot)khatiwada}
}
