% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_AnnualSumSC}
\alias{J2K_AnnualSumSC}
\title{Annual plot of}
\usage{
J2K_AnnualSumSC()
}
\arguments{
\item{TimeLoop}{Output from J2K model}
}
\value{
Annual input and output information
}
\description{
Visulization of annual input and outputs
}
\examples{
WatBal_AnnualSum_SC <- plot(x=years, y=value)

}
