% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_FixmyNG2G}
\alias{J2K_FixmyNG2G}
\title{Fixing the Non-Glacial to Glacial Routing issue}
\usage{
J2K_FixmyNG2G()
}
\arguments{
\item{hrus}{Considers the hrus for the J2K model}
}
\value{
Helps in solving the non-glacier to glacier routing issue and saving the improved file
}
\description{
This will help to idenify the HRU ID which were providing the water from the non-glacier area to the glacier area and also solve the issue and save the file.
}
\examples{
J2K_FixmyNG2G <- write.csv(improved file, in the working folder)

}
