% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_inVSoutInfo}
\alias{J2K_inVSoutInfo}
\title{Annual plot of input vs output}
\usage{
J2K_inVSoutInfo()
}
\arguments{
\item{TimeLoop}{Considers the timeloop file from the Output folder from J2K model}
}
\value{
Annual input and output. Note_1:It should have atleast 360 days to be considered a year. Note_2: The sum of the variables is considered for the visulization.
}
\description{
Visulization of annual input and outputs
}
\examples{
J2K_inVSoutInfo <- plot(x=input_output, y=Yearly_sum_value)

}
