% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_MonthMeanSC}
\alias{J2K_MonthMeanSC}
\title{Monthly snow cover area (km2)}
\usage{
J2K_MonthMeanSC()
}
\arguments{
\item{TimeLoop}{It considers the input as MODIS and the output from J2K model and makes a comparative plot}
}
\value{
Monthly snow cover area information
}
\description{
Helps in the visulization of the monthly average snow cover area. This will be particularly useful if someone is trying to observe the change in each month.
}
\examples{
J2K_MonthMeanSC <- plot(x=years, y=monthly area)

}
