% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_RainVsRunoffmmMonthly}
\alias{J2K_RainVsRunoffmmMonthly}
\title{Monthly precipitation and discharge ( mm)}
\usage{
J2K_RainVsRunoffmmMonthly()
}
\arguments{
\item{input}{Considers the input value for the J2K model}
}
\value{
Helps in the monthly visulization of the P vs Q  in same unit
}
\description{
This function helps in the visulization of the precipitation and discharge in the same unit(mm). Daily average value of the provided station/gauge point is computed to form the montly value of precipitaiton vs discharge
}
\examples{
J2K_RainVsRunoffmmMonthly <- plot(x=years, y=value (mm))

}
