% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_RainVsRunoffmmYearly}
\alias{J2K_RainVsRunoffmmYearly}
\title{Annual sum of precipitation and discharge in mm}
\usage{
J2K_RainVsRunoffmmYearly()
}
\arguments{
\item{input}{Considers the input value for the J2K model}
}
\value{
Helps in the annual visulization of the P vs Q  in same unit
}
\description{
This function helps in the visulization of annual of both the precipitation and discharge in the same unit(mm). Daily average precipitation from all the available station is computed to form the annual sum of the precipitaiton vs discharge of provided discharge station.
}
\examples{
J2K_RainVsRunoffmmYearly <- plot(x=years, y=value (mm))

}
