% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_snowcoverTS}
\alias{J2K_snowcoverTS}
\title{Annual plot of Annual Snow Cover}
\usage{
J2K_snowcoverTS()
}
\arguments{
\item{TimeLoop}{Output from J2K model}
}
\value{
Annual input and output information
}
\description{
Visulization of annual input and outputs
}
\examples{
WatBal_TS_snowcover <- plot(x=years, y=value)

}
