% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_WatBalplot}
\alias{J2K_WatBalplot}
\title{Daily water balance of the basin}
\usage{
J2K_WatBalplot()
}
\arguments{
\item{TimeLoop}{The timeloop file present in output folder of J2K model}
}
\value{
Plot with the daily water balance
}
\description{
More info in the basic of results on J2000 hydrological model
}
\examples{
WatBal_plot() <- plot(x=years, y= daily_water_storage_in_basin)

}
