% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_WatBalsummarySave}
\alias{J2K_WatBalsummarySave}
\title{Export the data}
\usage{
J2K_WatBalsummarySave()
}
\arguments{
\item{TimeLoop}{Based on the timeloop file the yearly value of the different variables and their percentage are saved as .csv file in the Output folder of the model}
}
\value{
Saves the summary of the waterbalance in the folder
}
\description{
Export the summary data in the working folder.The quick overview is displayed in the console. The supplimnet informaiton you would get is the percentage of the different variables. For example: yearly percentage of glacier in runoff. percentage of snowrunoff, evapotranspiration etc.
}
\examples{
J2K_WatBalsummarySave<- write.csv("Info_On_Water_Balance.csv" in workingFolder)

}
