% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{package}
\name{WatBal}
\alias{WatBal}
\title{analysisJ2Kmodel Package}
\description{
This package is create exclusively to test the water balance of the J2000 hydrological model. Currently it is in the beta mode. Comments are appreciated. Note: few considerations are to be made for running the code. 1) Make sure you have installed and loaded libraries of:"data.table", "ggplot2", "dplyr", "tidyr", "knitr". 2) As an initail set up, two input variables/paths are suggested (i) Output <- "The path where timeloop is situated (ii) Input <- "The path where the input variable are stored"
}
\author{
Kabiraj \email{kabiraj.khatiwada@gmail.com}
}
