% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\docType{package}
\name{analyzeJ2Kmodel}
\alias{analyzeJ2Kmodel}
\title{analyzeJ2Kmodel Package}
\description{
This package is created exclusively for testing the basics information while setting up the J2000 hydrological model in a new catchment. Currently it is in the beta mode, therefore, comments and suggestions are are highly appreciated.
**Note: For now few considerations are to be made for running the package. Firstly make sure you put the "workingFolder" <- "The path where the jam file is situated. Based on this all the calucations are made.
You can see the further information on "https://github.com/kabiraj404/analyzeJ2Kmodel/blob/master/README.md"
}
\author{
Kabiraj \email{kabiraj.khatiwada@gmail.com}
}
