% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_RainVsRunoffmmMonthly}
\alias{J2K_RainVsRunoffmmMonthly}
\title{Annual plot of Rain Vs Runoff in MonthlyScale}
\usage{
J2K_RainVsRunoffmmMonthly()
}
\arguments{
\item{input}{Output from J2K model}
}
\value{
Annual input and output information
}
\description{
Visulization of annual input and outputs
}
\examples{
WatBal_RainfallvsRunoff_M <- plot(x=years, y=value)

}
