% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_snowcoverTS}
\alias{J2K_snowcoverTS}
\title{Annual plot of Annual Snow Cover}
\usage{
J2K_snowcoverTS()
}
\arguments{
\item{TimeLoop}{Based on the timeLoop of J2K model}
}
\value{
visual of MODIS and model snow cover area daily
}
\description{
Helps in the visulization of the daily snow cover area
}
\examples{
J2K_snowcoverTS <- plot(x=days, y=snow cover area )

}
