% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_WatBalMajorComps}
\alias{J2K_WatBalMajorComps}
\title{Major components of the water balance}
\usage{
J2K_WatBalMajorComps()
}
\arguments{
\item{TimeLoop}{Output from J2K model}
}
\value{
Info on the water balance
}
\description{
Visulization of major components groups in Input, Output and Snow/Glacier
}
\examples{
WatBal_majorcomponents() <- plot(x=varables, y=value)

}
