% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_WatBalplot}
\alias{J2K_WatBalplot}
\title{Daily water balance of the basin}
\usage{
J2K_WatBalplot()
}
\arguments{
\item{TimeLoop}{The plot is based on the timeloop.dat file present in output folder of J2K model}
}
\value{
Plot with the daily water balance
}
\description{
Plots the daily storage in the basin. We know that the Input in a basin is equal to output _+ storage. Therefore the plot aims to show the storage of the basin. Theoritically a perfect model should have the storage(waterbalance) near to zero so that the water cycle is complete and there is no water water storage in the basin. If you see some spikes or higher values in the plot, it might be that the water has been stuck in some HRUs. Check them out.
}
\examples{
J2K_WatBalplot <- plot(x=years, y= daily_water_storage_in_basin)

}
