% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisJ2Kmodel.R
\name{J2K_WatBalsummarySave}
\alias{J2K_WatBalsummarySave}
\title{Export the data}
\usage{
J2K_WatBalsummarySave()
}
\arguments{
\item{TimeLoop}{Output from J2K model}
}
\value{
Saves the summary of the waterbalance in the folder
}
\description{
Export the summary data in the Output folder
}
\examples{
WatBal_writesummary<- write.csv("Info_On_Water_Balance.csv" in Output folder)

}
